requirements.txt
asn1crypto==0.24.0
azure-common==1.1.18
azure-storage-blob==1.5.0
azure-storage-common==1.4.0
boto3==1.9.130
botocore==1.12.130
cachetools==3.1.0
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
cryptography==2.6.1
docutils==0.14
future==0.17.1
google-api-python-client==1.7.8
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.3.0
httplib2==0.12.1
idna==2.8
ijson==2.3
jmespath==0.9.4
oauthlib==3.0.1
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycparser==2.19
pycryptodomex==3.8.1
pygsheets==2.0.1
PyJWT==1.7.1
pyOpenSSL==19.0.0
python-dateutil==2.8.0
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.2.0
rsa==4.0
s3transfer==0.2.0
six==1.12.0
snowflake-connector-python==1.7.9
uritemplate==3.0.0
urllib3==1.24.1
